Refactored Room node ops
